chore(scripts): move teleop to scripts
