+++ pyproject.toml
@@ -168,7 +168,7 @@ lerobot-find-port="lerobot.scripts.lerobot_find_port:main"
 lerobot-record="lerobot.record:main"
 lerobot-replay="lerobot.replay:main"
 lerobot-setup-motors="lerobot.setup_motors:main"
-lerobot-teleoperate="lerobot.teleoperate:main"
+lerobot-teleoperate="lerobot.scripts.lerobot_teleoperate:main"
 lerobot-eval="lerobot.scripts.eval:main"
 lerobot-train="lerobot.scripts.train:main"
 lerobot-dataset-viz="lerobot.scripts.lerobot_dataset_viz:main"
